fix: cast collection approver parameters

+++ backend/src/repositories/postgres/transactions.repository.js
@@ -190,7 +190,7 @@ async function updateCollectionTransaction(id, input) {
       category_id = $5,
       amount_minor = $6,
       occurred_on = $7,
-      approved_by_user_id = CASE WHEN $3::smallint = ${TX_STATUS.APPROVED}::smallint THEN $8 ELSE NULL END,
+      approved_by_user_id = CASE WHEN $3::smallint = ${TX_STATUS.APPROVED}::smallint THEN $8::integer ELSE NULL::integer END,
       approved_at = CASE WHEN $3::smallint = ${TX_STATUS.APPROVED}::smallint THEN COALESCE($9, NOW()) ELSE NULL END,
       note = $10,
       updated_at = NOW()
@@ -215,7 +215,7 @@ async function updateTransactionStatus(id, status, approvedByUserId) {
     `UPDATE transactions
      SET
       status = $2::smallint,
-      approved_by_user_id = CASE WHEN $2::smallint = ${TX_STATUS.APPROVED}::smallint THEN $3 ELSE NULL END,
+      approved_by_user_id = CASE WHEN $2::smallint = ${TX_STATUS.APPROVED}::smallint THEN $3::integer ELSE NULL::integer END,
       approved_at = CASE WHEN $2::smallint = ${TX_STATUS.APPROVED}::smallint THEN NOW() ELSE NULL END,
       updated_at = NOW()
      WHERE id = $1`,
